+++ yandex/resource_yandex_kubernetes_node_group.go
@@ -277,6 +277,21 @@ func resourceYandexKubernetesNodeGroup() *schema.Resource {
 								},
 							},
 						},
+						"gpu_settings": {
+							Type:     schema.TypeList,
+							Computed: true,
+							Optional: true,
+							MaxItems: 1,
+							Elem: &schema.Resource{
+								Schema: map[string]*schema.Schema{
+									"gpu_cluster_id": {
+										Type:     schema.TypeString,
+										Optional: true,
+										ForceNew: true,
+									},
+								},
+							},
+						},
 					},
 				},
 			},
@@ -826,7 +841,10 @@ func getNodeGroupTemplate(d *schema.ResourceData) (*k8s.NodeTemplate, error) {
 	if err != nil {
 		return nil, fmt.Errorf("error expanding container network while creating Kubernetes node group: %s", err)
 	}
-
+	gpuSettings, err := getNodeGroupGPUSettings(d)
+	if err != nil {
+		return nil, fmt.Errorf("error expanding gpu_settings while creating Kubernetes node group: %s", err)
+	}
 	tpl := &k8s.NodeTemplate{
 		PlatformId:               h.GetString("platform_id"),
 		ResourcesSpec:            getNodeGroupResourceSpec(d),
@@ -839,6 +857,7 @@ func getNodeGroupTemplate(d *schema.ResourceData) (*k8s.NodeTemplate, error) {
 		NetworkSettings:          ns,
 		ContainerRuntimeSettings: crs,
 		ContainerNetworkSettings: cns,
+		GpuSettings:              gpuSettings,
 		Name:                     h.GetString("name"),
 		Labels:                   labels,
 	}
@@ -980,6 +999,15 @@ func getNodeGroupContainerNetworkSettings(d *schema.ResourceData) (*k8s.NodeTemp
 		cns.SetPodMtu(int64(podMTU.(int)))
 	}
 	return cns, nil
+}
+
+func getNodeGroupGPUSettings(d *schema.ResourceData) (*k8s.GpuSettings, error) {
+	if gpuClusterID, ok := d.GetOk("instance_template.0.gpu_settings.0.gpu_cluster_id"); ok {
+		return &k8s.GpuSettings{
+			GpuClusterId: gpuClusterID.(string),
+		}, nil
+	}
+	return nil, nil
 }
 
 func flattenNodeGroupSchemaData(ng *k8s.NodeGroup, d *schema.ResourceData) error {
@@ -1228,6 +1256,7 @@ func flattenKubernetesNodeGroupTemplate(ngTpl *k8s.NodeTemplate) []map[string]in
 		"name":                      ngTpl.GetName(),
 		"labels":                    ngTpl.GetLabels(),
 		"container_network":         flattenKubernetesNodeGroupTemplateContainerNetwork(ngTpl.GetContainerNetworkSettings()),
+		"gpu_settings":              flattenKubernetesNodeGroupTemplateGPUSettings(ngTpl.GetGpuSettings()),
 	}
 
 	return []map[string]interface{}{tpl}
@@ -1395,6 +1424,17 @@ func flattenKubernetesNodeGroupTemplateContainerNetwork(p *k8s.NodeTemplate_Cont
 	return []map[string]interface{}{
 		{
 			"pod_mtu": int(p.GetPodMtu()),
+		},
+	}
+}
+
+func flattenKubernetesNodeGroupTemplateGPUSettings(p *k8s.GpuSettings) []map[string]interface{} {
+	if p == nil {
+		return []map[string]interface{}{}
+	}
+	return []map[string]interface{}{
+		{
+			"gpu_cluster_id": p.GetGpuClusterId(),
 		},
 	}
 }